+++ src/debugFunction.ts
@@ -1,5 +1,6 @@
-import { Keyboard } from "isaac-typescript-definitions";
+import { Dimension, Keyboard } from "isaac-typescript-definitions";
 import {
+  getDimension,
   isKeyboardPressed,
   log,
   printConsole,
@@ -24,6 +25,9 @@ export function init(): void {
 
 function debugCode(_params?: string) {
   // Add code here.
+  printConsole(
+    `GETTING HERE - CURRENT DIMENSION: ${Dimension[getDimension()]}`,
+  );
 }
 
 export function debugFunction(params?: string): void {