+++ src/features/changeCharOrder/callbacks/postRender.ts
@@ -89,6 +89,7 @@ function drawSeasonSprites() {
     );
     const posRender = Isaac.WorldToScreen(posButton);
     seasonSprite.RenderLayer(0, posRender);
+    Isaac.DebugString("GETTING HERE");
   }
 }