+++ src/features/optional/quality/showPills.ts
@@ -8,11 +8,14 @@ import {
 import {
   anyPlayerHasCollectible,
   getDefaultKColor,
+  getFalsePHDPillEffect,
   getNormalPillColorFromHorse,
   getNormalPillColors,
+  getPHDPillEffect,
   getPillEffectName,
   isActionPressedOnAnyInput,
   log,
+  NUM_PILLS_IN_POOL,
   saveDataManager,
 } from "isaacscript-common";
 import g from "../../../globals";
@@ -20,64 +23,19 @@ import { config } from "../../../modConfigMenu";
 import { initSprite } from "../../../sprite";
 import { PillDescription } from "../../../types/PillDescription";
 
-const NUM_PILLS_IN_POOL = 13;
+const FALSE_PHD_PILL_CONVERSIONS_RACING_PLUS: ReadonlyMap<
+  PillEffect,
+  PillEffect
+> = new Map([
+  // In vanilla, this converts to ???, but in Racing+ we manually convert it to I'm Excited!!!
+  [PillEffect.TELEPILLS, PillEffect.IM_EXCITED], // 19
 
-const PHD_PILL_CONVERSIONS: ReadonlyMap<PillEffect, PillEffect> = new Map([
-  [PillEffect.BAD_TRIP, PillEffect.BALLS_OF_STEEL], // 1
-  [PillEffect.HEALTH_DOWN, PillEffect.HEALTH_UP], // 6
-  [PillEffect.RANGE_DOWN, PillEffect.RANGE_UP], // 11
-  [PillEffect.SPEED_DOWN, PillEffect.SPEED_UP], // 13
-  [PillEffect.TEARS_DOWN, PillEffect.TEARS_UP], // 15
-  [PillEffect.LUCK_DOWN, PillEffect.LUCK_UP], // 17
-  [PillEffect.PARALYSIS, PillEffect.PHEROMONES], // 22
-  [PillEffect.AMNESIA, PillEffect.I_CAN_SEE_FOREVER], // 25
-  [PillEffect.R_U_A_WIZARD, PillEffect.POWER], // 27
-  [PillEffect.ADDICTED, PillEffect.PERCS], // 29
-  [PillEffect.QUESTION_MARKS, PillEffect.TELEPILLS], // 31
-  [PillEffect.RETRO_VISION, PillEffect.I_CAN_SEE_FOREVER], // 37
-  [PillEffect.X_LAX, PillEffect.SOMETHINGS_WRONG], // 39
-  [PillEffect.IM_EXCITED, PillEffect.IM_DROWSY], // 42
-  [PillEffect.HORF, PillEffect.GULP], // 44
-  [PillEffect.SHOT_SPEED_DOWN, PillEffect.SHOT_SPEED_UP], // 47
+  // In vanilla, this converts to Amnesia, but in Racing+ we manually convert it to Retro Vision.
+  [PillEffect.I_CAN_SEE_FOREVER, PillEffect.RETRO_VISION], // 23
+
+  // In vanilla, this converts to Amnesia, but in Racing+ we manually convert it to Horf!
+  [PillEffect.LEMON_PARTY, PillEffect.HORF], // 26
 ]);
-
-const FALSE_PHD_PILL_CONVERSIONS: ReadonlyMap<PillEffect, PillEffect> = new Map(
-  [
-    [PillEffect.BAD_GAS, PillEffect.HEALTH_DOWN], // 0
-    [PillEffect.BALLS_OF_STEEL, PillEffect.BAD_TRIP], // 2
-    [PillEffect.BOMBS_ARE_KEYS, PillEffect.TEARS_DOWN], // 3
-    [PillEffect.EXPLOSIVE_DIARRHEA, PillEffect.RANGE_DOWN], // 4
-    [PillEffect.FULL_HEALTH, PillEffect.BAD_TRIP], // 5
-    [PillEffect.HEALTH_UP, PillEffect.HEALTH_DOWN], // 7
-    [PillEffect.PRETTY_FLY, PillEffect.LUCK_DOWN], // 10
-    [PillEffect.RANGE_UP, PillEffect.RANGE_DOWN], // 12
-    [PillEffect.SPEED_UP, PillEffect.SPEED_DOWN], // 14
-    [PillEffect.TEARS_UP, PillEffect.TEARS_DOWN], // 16
-    [PillEffect.LUCK_UP, PillEffect.LUCK_DOWN], // 18
-    // In vanilla, this converts to ???, but in Racing+ we manually convert it to I'm Excited!!!
-    [PillEffect.TELEPILLS, PillEffect.IM_EXCITED], // 19
-    [PillEffect.FORTY_EIGHT_HOUR_ENERGY, PillEffect.SPEED_DOWN], // 20
-    [PillEffect.HEMATEMESIS, PillEffect.BAD_TRIP], // 21
-    // In vanilla, this converts to Amnesia, but in Racing+ we manually convert it to Retro Vision.
-    [PillEffect.I_CAN_SEE_FOREVER, PillEffect.RETRO_VISION], // 23
-    [PillEffect.PHEROMONES, PillEffect.PARALYSIS], // 24
-    // In vanilla, this converts to Amnesia, but in Racing+ we manually convert it to Horf!
-    [PillEffect.LEMON_PARTY, PillEffect.HORF], // 26
-    [PillEffect.PERCS, PillEffect.ADDICTED], // 28
-    [PillEffect.ONE_MAKES_YOU_LARGER, PillEffect.RANGE_DOWN], // 32
-    [PillEffect.ONE_MAKES_YOU_SMALL, PillEffect.SPEED_DOWN], // 33
-    [PillEffect.INFESTED_EXCLAMATION, PillEffect.TEARS_DOWN], // 34
-    [PillEffect.INFESTED_QUESTION, PillEffect.LUCK_DOWN], // 35
-    [PillEffect.POWER, PillEffect.R_U_A_WIZARD], // 36
-    [PillEffect.FRIENDS_TILL_THE_END, PillEffect.HEALTH_DOWN], // 38
-    [PillEffect.SOMETHINGS_WRONG, PillEffect.X_LAX], // 40
-    [PillEffect.IM_DROWSY, PillEffect.IM_EXCITED], // 41
-    [PillEffect.GULP, PillEffect.HORF], // 43
-    [PillEffect.FEELS_LIKE_IM_WALKING_ON_SUNSHINE, PillEffect.RETRO_VISION], // 45
-    [PillEffect.VURP, PillEffect.HORF], // 46
-    [PillEffect.SHOT_SPEED_UP, PillEffect.SHOT_SPEED_DOWN], // 48
-  ],
-);
 
 /** These are not meant to ever be reset. */
 const pillSprites = new Map<PillColor, Sprite>();
@@ -131,14 +89,10 @@ function checkPHD() {
   }
 
   v.run.PHD = true;
-  log("Converting bad pill effects.");
 
   // Change the text for any identified pills.
   for (const pillEntry of v.run.pillsIdentified) {
-    const newEffect = PHD_PILL_CONVERSIONS.get(pillEntry.effect);
-    if (newEffect !== undefined) {
-      pillEntry.effect = newEffect;
-    }
+    pillEntry.effect = getPHDPillEffect(pillEntry.effect);
   }
 }
 
@@ -157,11 +111,16 @@ function checkFalsePHD() {
 
   // Change the text for any identified pills.
   for (const pillEntry of v.run.pillsIdentified) {
-    const newEffect = FALSE_PHD_PILL_CONVERSIONS.get(pillEntry.effect);
-    if (newEffect !== undefined) {
-      pillEntry.effect = newEffect;
-    }
+    pillEntry.effect = getFalsePHDPillEffectRacingPlus(pillEntry.effect);
   }
+}
+
+/** Racing+ makes some modifications to the pill effects (since all curses are removed). */
+function getFalsePHDPillEffectRacingPlus(pillEffect: PillEffect): PillEffect {
+  const newPillEffect = getFalsePHDPillEffect(pillEffect);
+  const racingPlusEffect =
+    FALSE_PHD_PILL_CONVERSIONS_RACING_PLUS.get(newPillEffect);
+  return racingPlusEffect === undefined ? newPillEffect : racingPlusEffect;
 }
 
 // ModCallback.POST_RENDER (2)