+++ src/types/EntityLocation.ts
@@ -1,4 +1,0 @@
-export interface EntityLocation {
-  roomListIndex: int;
-  position: Vector;
-}